[Add] User Repository -> Next userusecase

+++ domain/model/base.go
@@ -4,7 +4,7 @@ import "time"
 
 // BaseModel is ...
 type BaseModel struct {
-	ID        uint       `gorm:"primary_key" json:"id"`
+	ID        int64      `gorm:"primary_key" json:"id"`
 	CreatedAt time.Time  `json:"-"`
 	UpdatedAt time.Time  `json:"-"`
 	DeletedAt *time.Time `sql:"index" json:"-" `
