+++ domain/model/shop.go
@@ -5,8 +5,8 @@ type Shop struct {
 	BaseModel
 	ImagePath    string
 	Name         string
-	Latitude     float64
-	Longitude    float64
+	Latitude     float32
+	Longitude    float32
 	Introduction string `gorm:"type:varchar(1000);" `
 	Vegetables   []Vegetable
 }
